refactor: migrate global types to configs
