+++ packages/core/src/config/index.ts
@@ -1,10 +1,3 @@
-import sanityClient from '@sanity/client';
-
-export const sanityConfig = {
-  projectId: 'gc3hakk3',
-  dataset: 'production',
-  apiVersion: 'v1',
-  useCdn: false, // `false` if you want to ensure fresh data
-};
-
-export const client = sanityClient(sanityConfig);
+export * from './languages';
+export * from './sanity';
+export * from './theme';
