+++ packages/core/src/index.ts
@@ -1,3 +1,3 @@
+export * from './config';
 export * from './domains';
-export * from './types';
 export * from './utils';
